// src/app/_interfaces/vg/vg.interface.ts
export interface InterfaceVG {
  token: string;
  idFuncionario: number;
  idFuncionarioFuncao: number;
  funcionario: string;
  idTerminal: number;
  terminal: string;
  idArmazem: number;
  armazem: string;
  idLoja: number;
  loja: string;
  localidadeLoja: string;
  idEmpresa: number;
  empresa: string;
  url: string;
  lingua: string;
  razaoSocial: string;
  corAgenda: string;
  corTexto: string;
  veterinaria: string;
  estetica: string;
  nutricao: string;
  cabeleireiro: string;
  audiologia: string;
  otica: string;
  idMeteo: string;
  idMensagem: string;
  timeoutOnline: number;
  medica: string;
  idLojaTamanho: number;
  idClienteTamanho: number;
  medicas: string;
  idTurno: number;
  ivaIncluido: number;
  acessoConfig: number;
  acessoChat: number;
  acessoFerramentas: number;
  software: number;
}